src/execution/position_sizer.py
"""
Trade Execution Optimization with Bayesian Optimization and Adaptive Risk Management.

Key features:
- Dynamic position sizing based on market impact prediction
- Bayesian optimization for parameter tuning
- Adaptive stop-loss and take-profit levels
- Risk-adjusted position scaling
"""

import os
import numpy as np
from scipy.stats import norm
from typing import Tuple, Dict
import pandas as pd
from dataclasses import dataclass
from sklearn.gaussian_process import GaussianProcessRegressor
from sklearn.gaussian_process.kernels import RBF, ConstantKernel
import sys

sys.path.append(
    os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
)
from src.utils.log_utils import setup_logging

logger = setup_logging(__name__)


@dataclass
class PositionConfig:
    """Configuration for position sizing."""

    base_position: float  # Base position size (α)
    min_position: float  # Minimum position size
    max_position: float  # Maximum position size
    vpin_threshold: float = 0.7  # Threshold for high VPIN
    vol_lookback: int = 21  # Days for volatility calculation
    risk_limit: float = 0.02  # Maximum risk per trade (2%)


class BayesianPositionSizer:
    """
    Dynamic position sizing using Bayesian optimization.

    The position size is determined by:
    Size_t = α * exp(β * VPIN_t) * (1/σ_t)

    where:
    - α: Base position size
    - β: VPIN sensitivity (optimized)
    - VPIN_t: Current VPIN value
    - σ_t: Current volatility
    """

    def __init__(self, config: PositionConfig):
        self.config = config

        # Initialize Gaussian Process for parameter optimization
        kernel = ConstantKernel(1.0) * RBF(length_scale=1.0)
        self.gp = GaussianProcessRegressor(
            kernel=kernel, n_restarts_optimizer=10, random_state=42
        )

        # Parameter bounds for optimization
        self.bounds = {
            "vpin_beta": (0.5, 2.0),  # VPIN sensitivity
            "vol_beta": (0.5, 2.0),  # Volatility sensitivity
        }

        # Historical performance tracking
        self.history = []

    def calculate_position_size(
        self, vpin: float, volatility: float, market_impact_pred: float
    ) -> float:
        """
        Calculate optimal position size based on current market conditions.

        Args:
            vpin: Current VPIN value
            volatility: Current volatility
            market_impact_pred: Predicted market impact

        Returns:
            float: Optimal position size
        """
        # Get current optimal parameters
        params = self._get_current_parameters()

        # Calculate base position size with VPIN and volatility adjustment
        position = self.config.base_position * np.exp(params["vpin_beta"] * vpin)
        position *= 1.0 / (volatility ** params["vol_beta"])

        # Adjust for predicted market impact
        impact_adjustment = 1.0 - abs(market_impact_pred)
        position *= impact_adjustment

        # Apply position limits
        position = np.clip(position, self.config.min_position, self.config.max_position)

        return position

    def calculate_risk_levels(
        self, price: float, volatility: float, vpin: float
    ) -> Tuple[float, float]:
        """
        Calculate adaptive stop-loss and take-profit levels.

        Args:
            price: Current price
            volatility: Current volatility
            vpin: Current VPIN value

        Returns:
            Tuple[float, float]: (stop_loss, take_profit) prices
        """
        # Base levels from volatility
        base_sl = 1.5 * volatility
        base_tp = 2.5 * volatility

        # Adjust based on VPIN
        if vpin > self.config.vpin_threshold:
            # Tighter stops during high VPIN periods
            sl_pct = max(0.5, base_sl * 0.8)
            tp_pct = min(1.0, base_tp * 1.2)
        else:
            sl_pct = max(0.5, base_sl)
            tp_pct = min(1.0, base_tp)

        stop_loss = price * (1 - sl_pct)
        take_profit = price * (1 + tp_pct)

        return stop_loss, take_profit

    def update_model(self, trade_result: Dict):
        """
        Update the Bayesian optimization model with trade results.

        Args:
            trade_result: Dictionary containing trade performance metrics
        """
        self.history.append(trade_result)

        if len(self.history) >= 10:  # Need minimum samples for optimization
            # Prepare training data
            X = np.array([[h["vpin_beta"], h["vol_beta"]] for h in self.history])
            y = np.array([h["pnl"] for h in self.history])

            # Fit GP model
            self.gp.fit(X, y)

            # Update parameters through optimization
            self._optimize_parameters()

    def _get_current_parameters(self) -> Dict[str, float]:
        """Get current optimal parameters."""
        if not self.history:
            return {"vpin_beta": 1.0, "vol_beta": 1.0}

        # Use GP prediction for optimal parameters
        param_space = np.array(
            [
                [1.0, 1.0],  # Default parameters
                [1.5, 1.0],  # VPIN-sensitive
                [1.0, 1.5],  # Volatility-sensitive
                [1.5, 1.5],  # Balanced sensitivity
            ]
        )

        predictions = self.gp.predict(param_space)
        best_idx = np.argmax(predictions)

        return {
            "vpin_beta": param_space[best_idx, 0],
            "vol_beta": param_space[best_idx, 1],
        }

    def _optimize_parameters(self):
        """Optimize model parameters using Bayesian optimization."""
        from scipy.optimize import minimize

        def objective(params):
            x = params.reshape(1, -1)
            mean, std = self.gp.predict(x, return_std=True)
            return -(mean + 0.1 * std)  # Include exploration bonus

        # Start from current best parameters
        current_params = self._get_current_parameters()
        x0 = np.array([current_params["vpin_beta"], current_params["vol_beta"]])

        # Optimize
        bounds = [
            (self.bounds["vpin_beta"][0], self.bounds["vpin_beta"][1]),
            (self.bounds["vol_beta"][0], self.bounds["vol_beta"][1]),
        ]

        result = minimize(objective, x0, bounds=bounds, method="L-BFGS-B")

        logger.info(f"Optimized parameters: {result.x}")
