requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
seaborn>=0.11.0
torch>=1.9.0
tqdm>=4.62.0
python-dotenv>=0.19.0
backtrader>=1.9.76.123
scipy>=1.7.0
statsmodels>=0.13.0

# Development dependencies
pytest>=6.0
black>=21.5b2
flake8>=3.9.0
mypy>=0.910
jupyter>=1.0.0
notebook>=6.4.0
ipywidgets>=7.6.0



